requirements.txt
pillow
fastapi
uvicorn
streamlit
requests
psutil
numpy
joblib
scikit-learn
pandas
nltk
pydantic